Use long for max TTL value, update config property name

+++ implementation/src/main/java/io/smallrye/jwt/auth/principal/DefaultJWTTokenParser.java
@@ -227,7 +227,7 @@ public class DefaultJWTTokenParser {
     }
 
     private void verifyTimeToLive(JWTAuthContextInfo authContextInfo, JwtClaims claimsSet) throws ParseException {
-        final Integer maxTimeToLiveSecs = authContextInfo.getMaxTimeToLiveSecs();
+        final Long maxTimeToLiveSecs = authContextInfo.getMaxTimeToLiveSecs();
 
         if (maxTimeToLiveSecs != null) {
             final NumericDate iat;
