+++ implementation/src/main/java/io/smallrye/jwt/auth/principal/JWTAuthContextInfo.java
@@ -28,7 +28,7 @@ public class JWTAuthContextInfo {
     private RSAPublicKey signerKey;
     private String issuedBy;
     private int expGracePeriodSecs = 60;
-    private Integer maxTimeToLiveSecs;
+    private Long maxTimeToLiveSecs;
     private String publicKeyLocation;
     private Integer jwksRefreshInterval;
     private String tokenHeader = "Authorization";
@@ -105,11 +105,11 @@ public class JWTAuthContextInfo {
         this.expGracePeriodSecs = expGracePeriodSecs;
     }
 
-    public Integer getMaxTimeToLiveSecs() {
+    public Long getMaxTimeToLiveSecs() {
         return maxTimeToLiveSecs;
     }
 
-    public void setMaxTimeToLiveSecs(Integer maxTimeToLiveSecs) {
+    public void setMaxTimeToLiveSecs(Long maxTimeToLiveSecs) {
         this.maxTimeToLiveSecs = maxTimeToLiveSecs;
     }
